# docs/swagger.yaml
basePath: /
definitions:
  controller.Customer:
    properties:
      _id:
        example: "1234567"
        type: string
      alamat:
        example: jl.sarijadi
        type: string
      deskripsi:
        example: nasi goreng dengan telor dan daging
        type: string
      email:
        example:
        - Xaviera_89@gmail.com
        - Putri_90@gmail.com
        items:
          type: string
        type: array
      gambar:
        example: https://i.pinimg.com/564x/94/82/ab/9482ab2e248d249e7daa7fd6924c8d3b.jpg
        type: string
      harga:
        example: 15000
        type: integer
      nama:
        example: xavieraa putri
        type: string
      nama_produk:
        example: Nasi Goreng
        type: string
      phone_number:
        example: "085798654096"
        type: string
      stok:
        example: "10"
        type: string
    type: object
  controller.ReqCustomer:
    properties:
      alamat:
        example: jl.sarijadi
        type: string
      deskripsi:
        example: nasi goreng dengan telor dan daging
        type: string
      email:
        example:
        - Xaviera_89@gmail.com
        - Putri_90@gmail.com
        items:
          type: string
        type: array
      gambar:
        example: https://i.pinimg.com/564x/94/82/ab/9482ab2e248d249e7daa7fd6924c8d3b.jpg
        type: string
      harga:
        example: 15000
        type: integer
      nama:
        example: Tes swager
        type: string
      nama_produk:
        example: Nasi Goreng
        type: string
      phone_number:
        example: "085798654096"
        type: string
      stok:
        example: "10"
        type: string
    type: object
host: ats-714220023-serlipariela-38bba14820aa.herokuapp.com
info:
  contact:
    email: 714220023@std.ulbi.ac.id
    name: API Support
    url: https://github.com/serlip06
  description: This is a sample swagger for Fiber
  title: TES SWAGGER ULBI
  version: "1.0"
paths:
  /customer:
    get:
      consumes:
      - application/json
      description: Mengambil semua data customer.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.Customer'
      summary: Get All Data Customer.
      tags:
      - Customer
  /customer/{id}:
    get:
      consumes:
      - application/json
      description: Ambil per ID data customer.
      parameters:
      - description: Masukan ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.Customer'
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Get By ID Data Customer.
      tags:
      - Customer
  /insert:
    post:
      consumes:
      - application/json
      description: Input data customer.
      parameters:
      - description: Payload Body [RAW]
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controller.ReqCustomer'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.Customer'
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: Insert data customer.
      tags:
      - Customer
schemes:
- https
- http
swagger: "2.0"